feat(ui): add side panel layout to Capabilities page
Move Capabilities page into (main) route group to get consistent layout
with other pages including the sidebar navigation.

Changes:
- Move page from /capabilities to (main)/capabilities
- Add Header component matching other pages
- Add 3-column grid layout with side panel
- Side panel includes: summary stats, categories, about info
- Main content area shows capability cards in 2-column grid
